+++ NEWS.md
@@ -5,6 +5,8 @@ Added kernels for factors.
 Changed default number of restarts to zero, and added checking more starting
 points. Should make it faster.
 
+Can give in data as data and formula instead of matrix and vector.
+
 # GauPro 0.2.4
 
 Very minor change in Rcpp code to remove CRAN error.